+++ its/samples/workspace-java.code-workspace
@@ -1,7 +1,7 @@
 {
 	"folders": [
 		{
-			"path": "/home/julien/Prog/Projects/sonarlint/sonarlint-vscode/its/samples/sample-java-maven-multi-module"
+			"path": "./sample-java-maven-multi-module"
 		}
 	]
 }
