+++ tsconfig.json
@@ -12,6 +12,7 @@
     },
     "exclude": [
         "node_modules",
-        ".vscode-test"
+        ".vscode-test",
+        "its"
     ]
 }
